requirements.txt
aiokafka==0.8.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.11.0
async-timeout==5.0.1
black==25.11.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==4.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
colorama==0.4.6
crashtest==0.4.1
cryptography==45.0.3
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
Django==5.2.9
djangorestframework==3.16.1
dulwich==0.22.8
et_xmlfile==2.0.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flake8==7.2.0
google-api-core==2.24.2
google-api-python-client==2.108.0
google-auth==2.38.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.69.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.2
idna==3.11
iniconfig==2.1.0
installer==0.7.0
isort==7.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.4
oauthlib==3.2.2
openai==1.3.7
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pandas-stubs==2.2.3.250308
pathspec==0.12.1
pbs-installer==2025.3.17
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
proto-plus==1.26.1
protobuf==6.30.2
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyflakes==3.3.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytokens==0.3.0
pytz==2025.2
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlparse==0.5.4
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
trove-classifiers==2025.3.19.19
types-pytz==2025.2.0.20250326
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.6.2
virtualenv==20.30.0
zstandard==0.23.0